Optimize Docker image build time durign development
During the development, the image may be rebuilt very often. And, as
the build time must be as short as possible, the following changes are
introduced:
1. The Dockerfile is rewritten in a way when the most cacheable
instructions are placed first. "cacheable" means that a change to a
source file will cause the build instead of using a cached layer (for
instance, `requirements.txt` will be updated less often than
`thenumbers` package; however, `entry.sh` will be even more stable than
`requirements.txt`).
2. `build_dev_image` command is added which uses cache during building
of a Docker image.