Show help when 0 arguments pass to `workflows` script
The `workflows` script always requires at least one argument in order to
work. If it's run without an argument, it just stops. That's why from a
usability perspective it is good to print a help message if there are no
arguments. And exit with exit code 1 also as this prevents the incorrect
script usage in a pipeline configuration or scripting.